gemini/pyproject.toml
[project]
name = "gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.12.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
